Build Linux packages for multiple architectures
Currently, when building a Linux package on e.g. an ARM64 machine, the
package produced will be marked as an ARM64 package (because of the
"-a native" flag passed to fpm), but will contain an AMD64
executable (because the build-linux target sets GOARCH
unconditionally).

With this change, the desired architecture can be changed by setting
the LINUX_PACKAGE_GOARCH variable, so that the package will always
match the binary it contains. Also, the "packages" target builds
packages for both AMD64 and ARM64.